Add JS API for compiling-to-JavaScript functionality.
