+++ index.js
@@ -1,5 +1,6 @@
 var schema = require('protocol-buffers-schema')
 var compile = require('./compile')
+var compiletojs = require('./compiletojs')
 
 var flatten = function (values) {
   if (!values) return null
@@ -34,4 +35,8 @@ module.exports = function (proto, opts) {
   }
 
   return new Messages()
+}
+
+module.exports.toJS = function (proto) {
+  return compiletojs(module.exports(proto))
 }
