+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "protocol-buffers",
-  "version": "4.0.4",
+  "version": "4.1.0",
   "description": "Protocol Buffers for Node.js",
   "repository": {
     "type": "git",
